+++ Controller.cpp
@@ -57,7 +57,7 @@ void Controller::changeHistoricalOrder(bool cronologicalOrder, QString txt, QDat
 
 void Controller::doTransaction(bool d, std::map<std::string, QString> dataTransaction, QDate date) {
 
-    Transaction transaction(d, dataTransaction["payerName"], dataTransaction["payerIBAN"], dataTransaction["receiverName"], dataTransaction["receiverIBAN"], (dataTransaction["amount"]).toInt(), dataTransaction["causal"], date );
+    Transaction transaction(d, dataTransaction["payerName"], dataTransaction["payerIBAN"], dataTransaction["receiverName"], dataTransaction["receiverIBAN"], (dataTransaction["amount"]).toFloat(), dataTransaction["causal"], date );
 
     switch(model->doTransaction(transaction)) {
         case 0: model->makeMessageDialogNoButtons("transazione riuscita", "Transazione effettuata con successo!");
